+++ src/components/ClubRegister/index.ts
@@ -1,0 +1,2 @@
+export * from './ClubRegisterForm';
+export * from './ClubImageUpload';
