--- src/pages/admin/auth/register/ModifyClubRegisterPage.tsx
@@ -1,5 +1,0 @@
-const ModifyClubRegisterPage = () => {
-    return <div>동아리 어드민 등록 정보 수정 페이지</div>;
-};
-
-export { ModifyClubRegisterPage };
